Move .extract() tests into separated folder
