+++ test/extract/01-string-as-model.model.json
