+++ test/extract/01-string-as-model.result.json
