+++ test/extract/01-string-as-model.source.html
