+++ test/extract/02-string-in-array-model.model.json
