+++ test/extract/02-string-in-array-model.result.json
