+++ test/extract/02-string-in-array-model.source.html
