+++ test/extract/02-string-in-array-model.test.js
@@ -10,9 +10,9 @@ let model = require(path.join(__dirname, testName + '.model'))
 let expected = require(path.join(__dirname, testName + '.result'))
 
 let test = require('tap').test
-let scrapy = require('..')
+let scrapy = require('../..')
 
-test('should process single-string models as selector and return a string containing its textContent', t => {
+test('should process single-array with single-string as a multi-selection query and return array of textContent', t => {
   let result = scrapy.extract(source, model)
   t.deepEqual(result, expected)
   t.end()
