+++ test/extract/03-single-level-object-model.model.json
