+++ test/extract/03-single-level-object-model.test.js
@@ -10,9 +10,9 @@ let model = require(path.join(__dirname, testName + '.model'))
 let expected = require(path.join(__dirname, testName + '.result'))
 
 let test = require('tap').test
-let scrapy = require('..')
+let scrapy = require('../..')
 
-test('should process single-array with single-string as a multi-selection query and return array of textContent', t => {
+test('should process object values as selectors and return object with same structure, containing testContent of each selector', t => {
   let result = scrapy.extract(source, model)
   t.deepEqual(result, expected)
   t.end()
