+++ test/extract/05-multi-level-object-model-with-array.model.json
