+++ test/extract/05-multi-level-object-model-with-array.result.json
