+++ test/extract/06-text-content.test.js
@@ -10,9 +10,9 @@ let model = require(path.join(__dirname, testName + '.model'))
 let expected = require(path.join(__dirname, testName + '.result'))
 
 let test = require('tap').test
-let scrapy = require('..')
+let scrapy = require('../..')
 
-test('empty nodes and nodes without text children should not break getTextContent', t => {
+test('textContent should include children\'s textContent, without the actual tags', t => {
   let result = scrapy.extract(source, model)
   t.deepEqual(result, expected)
   t.end()
