+++ test/extract/07-text-content-empty-node.source.html
