+++ test/extract/08-array-of-objects.model.json
