+++ test/extract/08-array-of-objects.result.json
