+++ test/extract/08-array-of-objects.source.html
