+++ test/extract/09-array-of-arrays.model.json
