+++ test/extract/10-array-of-arrays-and-objects.model.json
